apps/backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg[binary]==3.2.3
celery[redis]==5.4.0
numpy==2.1.3
scikit-learn==1.5.2
xgboost==2.1.2
joblib==1.4.2
